+++ database/migrations/2016_06_01_000005_create_oauth_personal_access_clients_table.php
@@ -13,11 +13,13 @@ class CreateOauthPersonalAccessClientsTable extends Migration
      */
     public function up()
     {
-        Schema::create('oauth_personal_access_clients', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('client_id')->index();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('oauth_personal_access_clients')) {
+            Schema::create('oauth_personal_access_clients', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('client_id')->index();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,6 +29,8 @@ class CreateOauthPersonalAccessClientsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('oauth_personal_access_clients');
+        if (Schema::hasTable('oauth_personal_access_clients')) {
+            Schema::dropIfExists('oauth_personal_access_clients');
+        }
     }
 }
